inner utils and extract slice-atom-array
