+++ src/index.tsx
@@ -4,3 +4,4 @@ export { default as focusAtom } from './focus-atom'
 export { default as localStorageAtom } from './localstorage-atom'
 export * from './types'
 export * from './slice-atom-object'
+export * from './slice-atom-array'
